[#430] Refactor to a trait
