+++ modules/apigee_edge_teams/src/Form/TeamAppCredentialRevokeForm.php
@@ -19,54 +19,13 @@
 
 namespace Drupal\apigee_edge_teams\Form;
 
-use Drupal\apigee_edge\Entity\Controller\AppCredentialControllerInterface;
 use Drupal\apigee_edge\Form\AppCredentialRevokeFormBase;
-use Drupal\apigee_edge_teams\Entity\Controller\TeamAppCredentialControllerFactoryInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Provides revoke confirmation form for team app credential.
  */
 class TeamAppCredentialRevokeForm extends AppCredentialRevokeFormBase {
 
-  /**
-   * The app credential controller factory.
-   *
-   * @var \Drupal\apigee_edge_teams\Entity\Controller\TeamAppCredentialControllerFactoryInterface
-   */
-  protected $appCredentialControllerFactory;
-
-  /**
-   * TeamAppDeleteCredentialForm constructor.
-   *
-   * @param \Drupal\apigee_edge_teams\Entity\Controller\TeamAppCredentialControllerFactoryInterface $app_credential_controller_factory
-   *   The app credential controller factory.
-   */
-  public function __construct(TeamAppCredentialControllerFactoryInterface $app_credential_controller_factory) {
-    $this->appCredentialControllerFactory = $app_credential_controller_factory;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('apigee_edge_teams.controller.team_app_credential_controller_factory')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function appCredentialController(string $owner, string $app_name): AppCredentialControllerInterface {
-    return $this->appCredentialControllerFactory->teamAppCredentialController($owner, $app_name);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getCancelUrl() {
-    return $this->app->toUrl();
-  }
+  use TeamAppCredentialFormTrait;
 
 }
