+++ src/Form/DeveloperAppCredentialApproveForm.php
@@ -19,53 +19,11 @@
 
 namespace Drupal\apigee_edge\Form;
 
-use Drupal\apigee_edge\Entity\Controller\AppCredentialControllerInterface;
-use Drupal\apigee_edge\Entity\Controller\DeveloperAppCredentialControllerFactoryInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
 /**
  * Provides approve confirmation form for developer app credential.
  */
 class DeveloperAppCredentialApproveForm extends AppCredentialApproveFormBase {
 
-  /**
-   * The app credential controller factory.
-   *
-   * @var \Drupal\apigee_edge\Entity\Controller\DeveloperAppCredentialControllerFactoryInterface
-   */
-  protected $appCredentialControllerFactory;
-
-  /**
-   * DeveloperAppDeleteCredentialForm constructor.
-   *
-   * @param \Drupal\apigee_edge\Entity\Controller\DeveloperAppCredentialControllerFactoryInterface $app_credential_controller_factory
-   *   The app credential controller factory.
-   */
-  public function __construct(DeveloperAppCredentialControllerFactoryInterface $app_credential_controller_factory) {
-    $this->appCredentialControllerFactory = $app_credential_controller_factory;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('apigee_edge.controller.developer_app_credential_factory')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function appCredentialController(string $owner, string $app_name): AppCredentialControllerInterface {
-    return $this->appCredentialControllerFactory->developerAppCredentialController($owner, $app_name);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getCancelUrl() {
-    return $this->app->toUrl('canonical-by-developer');
-  }
+  use DeveloperAppCredentialFormTrait;
 
 }
